Move navigation widgets to navigation/view
